+++ pkg/backup/create.go
@@ -136,7 +136,7 @@ func (b *Backuper) CreateBackup(backupName, tablePattern string, partitions []st
 	for _, disk := range disks {
 		diskMap[disk.Name] = disk.Path
 	}
-	partitionsToBackupMap, partitions := filesystemhelper.CreatePartitionsToBackupMap(b.ch, tables, nil, partitions)
+	partitionsToBackupMap, partitions := filesystemhelper.CreatePartitionsToBackupMap(ctx, b.ch, tables, nil, partitions)
 	// create
 	if b.cfg.ClickHouse.UseEmbeddedBackupRestore {
 		err = b.createBackupEmbedded(ctx, backupName, tablePattern, partitions, partitionsToBackupMap, schemaOnly, rbacOnly, configsOnly, tables, allDatabases, allFunctions, disks, diskMap, log, startBackup, version)
@@ -326,7 +326,9 @@ func (b *Backuper) createBackupEmbedded(ctx context.Context, backupName, tablePa
 	if len(backupResult) != 1 || (backupResult[0].Status != "BACKUP_COMPLETE" && backupResult[0].Status != "BACKUP_CREATED") {
 		return fmt.Errorf("backup return wrong results: %+v", backupResult)
 	}
-	backupDataSize := make([]uint64, 0)
+	backupDataSize := make([]struct {
+		Size uint64 `ch:"backup_data_size"`
+	}, 0)
 	if !schemaOnly {
 		if backupResult[0].CompressedSize == 0 {
 			chVersion, err := b.ch.GetVersion(ctx)
@@ -341,10 +343,14 @@ func (b *Backuper) createBackupEmbedded(ctx context.Context, backupName, tablePa
 				return err
 			}
 		} else {
-			backupDataSize = append(backupDataSize, backupResult[0].CompressedSize)
+			backupDataSize = append(backupDataSize, struct {
+				Size uint64 `ch:"backup_data_size"`
+			}{Size: backupResult[0].CompressedSize})
 		}
 	} else {
-		backupDataSize = append(backupDataSize, 0)
+		backupDataSize = append(backupDataSize, struct {
+			Size uint64 `ch:"backup_data_size"`
+		}{Size: 0})
 	}
 
 	log.Debug("calculate parts list from embedded backup disk")
@@ -382,7 +388,7 @@ func (b *Backuper) createBackupEmbedded(ctx context.Context, backupName, tablePa
 		}
 	}
 	backupMetaFile := path.Join(diskMap[b.cfg.ClickHouse.EmbeddedBackupDisk], backupName, "metadata.json")
-	if err := b.createBackupMetadata(ctx, backupMetaFile, backupName, backupVersion, "embedded", diskMap, disks, backupDataSize[0], backupMetadataSize, 0, 0, tableMetas, allDatabases, allFunctions, log); err != nil {
+	if err := b.createBackupMetadata(ctx, backupMetaFile, backupName, backupVersion, "embedded", diskMap, disks, backupDataSize[0].Size, backupMetadataSize, 0, 0, tableMetas, allDatabases, allFunctions, log); err != nil {
 		return err
 	}
 
@@ -486,7 +492,6 @@ func (b *Backuper) AddTableToBackup(ctx context.Context, backupName, shadowBacku
 		log.WithField("engine", table.Engine).Warnf("supports only schema backup")
 		return nil, nil, nil
 	}
-	//
 	if b.cfg.ClickHouse.CheckPartsColumns {
 		if err := b.ch.CheckSystemPartsColumns(ctx, table); err != nil {
 			return nil, nil, err
@@ -536,7 +541,7 @@ func (b *Backuper) AddTableToBackup(ctx context.Context, backupName, shadowBacku
 	}
 	// Unfreeze to unlock data on S3 disks, https://github.com/Altinity/clickhouse-backup/issues/423
 	if version > 21004000 {
-		if _, err := b.ch.QueryContext(ctx, fmt.Sprintf("ALTER TABLE `%s`.`%s` UNFREEZE WITH NAME '%s'", table.Database, table.Name, shadowBackupUUID)); err != nil {
+		if err := b.ch.QueryContext(ctx, fmt.Sprintf("ALTER TABLE `%s`.`%s` UNFREEZE WITH NAME '%s'", table.Database, table.Name, shadowBackupUUID)); err != nil {
 			return disksToPartsMap, realSize, err
 		}
 	}