+++ pkg/clickhouse/utils.go
@@ -1,41 +1,8 @@
 package clickhouse
 
 import (
-	"fmt"
-	"reflect"
 	"strings"
-
-	"github.com/jmoiron/sqlx/reflectx"
 )
-
-func baseType(t reflect.Type, expected reflect.Kind) (reflect.Type, error) {
-	t = reflectx.Deref(t)
-	if t.Kind() != expected {
-		return nil, fmt.Errorf("expected %s but got %s", expected, t.Kind())
-	}
-	return t, nil
-}
-
-func fieldsByTraversal(v reflect.Value, traversals [][]int, values []interface{}, ptrs bool) error {
-	v = reflect.Indirect(v)
-	if v.Kind() != reflect.Struct {
-		return fmt.Errorf("argument not a struct")
-	}
-
-	for i, traversal := range traversals {
-		if len(traversal) == 0 {
-			values[i] = new(interface{})
-			continue
-		}
-		f := reflectx.FieldByIndexes(v, traversal)
-		if ptrs {
-			values[i] = f.Addr().Interface()
-		} else {
-			values[i] = f.Interface()
-		}
-	}
-	return nil
-}
 
 func getDisksByPath(disks []Disk, dataPath string) []string {
 	resultDisks := make([]Disk, 0)
@@ -71,4 +38,12 @@ func GetDisksByPaths(disks []Disk, dataPaths []string) map[string]string {
 		}
 	}
 	return result
+}
+
+func ConvertToSlice(databases []Database) []string {
+	res := make([]string, len(databases))
+	for idx, dbName := range databases {
+		res[idx] = dbName.Name
+	}
+	return res
 }